jobmatcher_ai/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pytesseract
Pillow
sentence-transformers
